feat(share functionality): share developer's profile
-add share intent on details activity
add onclick method to share button to start share intent

[Finishes #156008359]
