--- app/src/main/java/com/example/ivanpaulrutale/convergelevelup/ListItem.java
@@ -1,15 +1,0 @@
-package com.example.ivanpaulrutale.convergelevelup;
-
-public class ListItem {
-    public ListItem(String username) {
-        this.username = username;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    private String username;
-
-
-}
